requirements.txt
PyQt6
opencv-python
imagehash
Pillow
librosa
numpy
send2trash
librosa
matplotlib